examples/chainflow-login/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
jinja2>=3.0.1
aiohttp>=3.7.4
pydantic>=1.8.2
# Hyperledger Indy bağımlılıkları - gerektiğinde yorum satırından çıkarın
# python3-indy>=1.16.0
